requirements.txt
laion_clap
git+https://github.com/openai/CLIP.git
tqdm
gradio
matplotlib
torch_mir_eval==0.4
opencv-python
librosa==0.9.2
transformers==4.30.2
moviepy==1.0.3
spacy
pytube 
ffmpeg-python